split actions file
