+++ src/UI/Components/Diagram/actions/editMode.test.ts
@@ -1,405 +1,19 @@
 import { dia } from "@inmanta/rappid";
 import { EmbeddedEntity, InstanceAttributeModel, ServiceModel } from "@/Core";
 import { InstanceWithRelations } from "@/Data/Managers/V2/GETTERS/GetInstanceWithRelations";
+import { containerModel, mockedInstanceWithRelations } from "../Mocks";
+import services from "../Mocks/services.json";
 import {
-  CreateModifierHandler,
-  FieldCreator,
-  createFormState,
-} from "../ServiceInstanceForm";
-import {
-  childModel,
-  containerModel,
-  mockedInstanceWithRelations,
-  parentModel,
-} from "./Mocks";
-import services from "./Mocks/services.json";
-import {
-  addDefaultEntities,
   appendEmbeddedEntity,
   appendInstance,
-  createComposerEntity,
   addInfoIcon,
-  populateGraphWithDefault,
-  updateAttributes,
-} from "./actions";
-import { ComposerPaper } from "./paper";
-import { ServiceEntityBlock } from "./shapes";
-import { defineObjectsForJointJS } from "./testSetup";
+} from "../actions/editMode";
+import { ComposerPaper } from "../paper";
+import { ServiceEntityBlock } from "../shapes";
+import { defineObjectsForJointJS } from "../testSetup";
 
 beforeAll(() => {
   defineObjectsForJointJS();
-});
-
-describe("createComposerEntity", () => {
-  it("creates a new core entity", () => {
-    const coreEntity = createComposerEntity({
-      serviceModel: parentModel,
-      isCore: true,
-      isInEditMode: false,
-    });
-
-    expect(coreEntity.get("holderName")).toBe(undefined);
-    expect(coreEntity.get("isEmbedded")).toBe(undefined);
-    expect(coreEntity.get("isCore")).toBe(true);
-    expect(coreEntity.attr("header/fill")).toBe(
-      "var(--pf-v5-global--palette--gold-400)",
-    );
-    expect(coreEntity.get("isInEditMode")).toBe(false);
-    expect(coreEntity.get("items")).toStrictEqual([[], []]);
-  });
-
-  it("creates a new embedded entity", () => {
-    const embeddedEntity = createComposerEntity({
-      serviceModel: containerModel.embedded_entities[0],
-      isCore: false,
-      isInEditMode: false,
-      isEmbedded: true,
-      holderName: containerModel.name,
-    });
-
-    expect(embeddedEntity.get("holderName")).toBe(containerModel.name);
-    expect(embeddedEntity.get("isEmbedded")).toBe(true);
-    expect(embeddedEntity.get("isCore")).toBe(undefined);
-    expect(embeddedEntity.attr("header/fill")).toBe(
-      "var(--pf-v5-global--palette--blue-400)",
-    );
-    expect(embeddedEntity.get("isInEditMode")).toBe(false);
-  });
-
-  it("creates a new entity with inster-service relations", () => {
-    const childEntity = createComposerEntity({
-      serviceModel: childModel,
-      isCore: false,
-      isInEditMode: false,
-    });
-
-    expect(childEntity.get("holderName")).toBe(undefined);
-    expect(childEntity.get("isEmbedded")).toBe(undefined);
-    expect(childEntity.get("isCore")).toBe(undefined);
-    expect(childEntity.get("isInEditMode")).toBe(false);
-    expect(childEntity.attr("header/fill")).toBe(
-      "var(--pf-v5-global--palette--purple-500)",
-    );
-    expect(childEntity.get("relatedTo")).toMatchObject(new Map());
-  });
-
-  it.each`
-    serviceModel                           | isCore   | isEmbedded
-    ${parentModel}                         | ${false} | ${false}
-    ${containerModel}                      | ${true}  | ${false}
-    ${containerModel.embedded_entities[0]} | ${false} | ${true}
-  `(
-    "it appends correctly attributes to the body",
-    ({ serviceModel, isCore, isEmbedded }) => {
-      const fieldCreator = new FieldCreator(new CreateModifierHandler());
-      const fields = fieldCreator.attributesToFields(serviceModel.attributes);
-      const attributes = createFormState(fields);
-
-      const coreEntity = createComposerEntity({
-        serviceModel,
-        isCore,
-        isEmbedded,
-        isInEditMode: false,
-        attributes,
-      });
-
-      expect(coreEntity.get("items")).toStrictEqual([
-        [
-          {
-            id: "name",
-            label: "name",
-            span: 2,
-          },
-        ],
-        [
-          {
-            id: "name_value",
-            label: "",
-          },
-        ],
-      ]);
-    },
-  );
-});
-
-describe("updateAttributes", () => {
-  it("set attributes, sanitizedAttrs and displayed items on initial update", () => {
-    const instanceAsTable = new ServiceEntityBlock().setName(parentModel.name);
-    const attributes = mockedInstanceWithRelations.instance
-      .active_attributes as InstanceAttributeModel; // instance based on parent-service model
-    const isInitial = true;
-
-    updateAttributes(
-      instanceAsTable,
-      ["name", "service_id"],
-      attributes,
-      isInitial,
-    );
-
-    expect(instanceAsTable.get("sanitizedAttrs")).toMatchObject({
-      name: "test12345",
-      service_id: "123412",
-      should_deploy_fail: false,
-    });
-    expect(instanceAsTable.get("instanceAttributes")).toMatchObject({
-      name: "test12345",
-      service_id: "123412",
-      should_deploy_fail: false,
-    });
-    expect(instanceAsTable.get("items")).toMatchObject([
-      [
-        {
-          id: "name",
-          label: "name",
-          span: 2,
-        },
-        {
-          id: "service_id",
-          label: "service_id",
-          span: 2,
-        },
-      ],
-      [
-        {
-          id: "name_value",
-          label: "test12345",
-        },
-        {
-          id: "service_id_value",
-          label: "123412",
-        },
-      ],
-    ]);
-  });
-
-  it("when there is no key attributes only attributes and sanitizedAttrs are set", () => {
-    const instanceAsTable = new ServiceEntityBlock().setName(parentModel.name);
-    const attributes = mockedInstanceWithRelations.instance
-      .active_attributes as InstanceAttributeModel; // instance based on parent-service model
-    const isInitial = true;
-
-    updateAttributes(instanceAsTable, [], attributes, isInitial);
-
-    expect(instanceAsTable.get("sanitizedAttrs")).toMatchObject({
-      name: "test12345",
-      service_id: "123412",
-      should_deploy_fail: false,
-    });
-    expect(instanceAsTable.get("instanceAttributes")).toMatchObject({
-      name: "test12345",
-      service_id: "123412",
-      should_deploy_fail: false,
-    });
-    expect(instanceAsTable.get("items")).toMatchObject([[], []]);
-  });
-
-  it("sanitized Attributes won't be overridden if isInitial property is set to false or if there are sanitizedAttributes already set", () => {
-    //sanitizedAttrs property is updated from the sidebar level as it requires fields to be present
-    const instanceAsTable = new ServiceEntityBlock().setName(parentModel.name);
-    const attributes = mockedInstanceWithRelations.instance
-      .active_attributes as InstanceAttributeModel; // instance based on parent-service model
-    const isInitial = true;
-
-    updateAttributes(instanceAsTable, [], attributes, isInitial);
-
-    expect(instanceAsTable.get("sanitizedAttrs")).toMatchObject({
-      name: "test12345",
-      service_id: "123412",
-      should_deploy_fail: false,
-    });
-    expect(instanceAsTable.get("instanceAttributes")).toMatchObject({
-      name: "test12345",
-      service_id: "123412",
-      should_deploy_fail: false,
-    });
-    expect(instanceAsTable.get("items")).toMatchObject([[], []]);
-
-    const updatedIsInitial = false;
-    const updatedAttributes = {
-      name: "newName",
-      service_id: "newId",
-      should_deploy_fail: false,
-    };
-
-    updateAttributes(instanceAsTable, [], updatedAttributes, updatedIsInitial);
-
-    expect(instanceAsTable.get("sanitizedAttrs")).toMatchObject({
-      name: "test12345",
-      service_id: "123412",
-      should_deploy_fail: false,
-    });
-    expect(instanceAsTable.get("instanceAttributes")).toMatchObject({
-      name: "newName",
-      service_id: "newId",
-      should_deploy_fail: false,
-    });
-
-    const updatedIsInitial2 = true;
-
-    updateAttributes(instanceAsTable, [], updatedAttributes, updatedIsInitial2);
-
-    expect(instanceAsTable.get("sanitizedAttrs")).toMatchObject({
-      name: "test12345",
-      service_id: "123412",
-      should_deploy_fail: false,
-    });
-    expect(instanceAsTable.get("instanceAttributes")).toMatchObject({
-      name: "newName",
-      service_id: "newId",
-      should_deploy_fail: false,
-    });
-  });
-});
-
-describe("addDefaultEntities", () => {
-  it("return empty array for service without embedded entities to add to the graph ", () => {
-    const graph = new dia.Graph({});
-    const embedded = addDefaultEntities(graph, parentModel);
-
-    expect(embedded).toMatchObject([]);
-  });
-
-  it("adds default entity for service with embedded entities to the graph ", () => {
-    const dispatchEventSpy = jest.spyOn(document, "dispatchEvent");
-    const graph = new dia.Graph({});
-
-    const embedded = addDefaultEntities(graph, containerModel);
-
-    expect(dispatchEventSpy).toHaveBeenCalledTimes(2);
-
-    //assert the arguments of the first call - calls is array of the arguments of each call
-    expect(
-      (dispatchEventSpy.mock.calls[0][0] as CustomEvent).detail,
-    ).toMatchObject({
-      name: "child_container",
-      id: expect.any(String),
-      relations: [{ current: 0, min: 1, name: "parent-service" }],
-    }); //add relations to Tracker
-
-    //assert the arguments of the second call - calls is array of the arguments of each call
-    expect(
-      (dispatchEventSpy.mock.calls[1][0] as CustomEvent).detail,
-    ).toMatchObject({
-      action: "add",
-      name: "child_container",
-    }); //update stencil state
-
-    expect(embedded.length).toBe(1);
-
-    expect(embedded[0].getName()).toStrictEqual("child_container");
-  });
-
-  it("adds all default entities to the graph for service with embedded entity with lower_limit > 1 ", () => {
-    const dispatchEventSpy = jest.spyOn(document, "dispatchEvent");
-    const graph = new dia.Graph({});
-    const customModel = {
-      ...containerModel,
-      embedded_entities: [
-        { ...containerModel.embedded_entities[0], lower_limit: 2 },
-      ],
-    };
-
-    const embedded = addDefaultEntities(graph, customModel);
-
-    expect(dispatchEventSpy).toHaveBeenCalledTimes(4);
-    expect(embedded.length).toBe(2);
-
-    expect(embedded[0].getName()).toStrictEqual("child_container");
-    expect(embedded[1].getName()).toStrictEqual("child_container");
-  });
-
-  it("adds default entity for service with nested embedded entities to the graph ", () => {
-    const dispatchEventSpy = jest.spyOn(document, "dispatchEvent");
-
-    const graph = new dia.Graph({});
-    const attributes = {
-      name: "",
-    };
-
-    const createdEntity1 = createComposerEntity({
-      serviceModel: {
-        ...containerModel.embedded_entities[0],
-        embedded_entities: [{ ...containerModel.embedded_entities[0] }],
-      },
-      isCore: false,
-      isInEditMode: false,
-      attributes,
-      isEmbedded: true,
-      holderName: "container-service",
-    });
-
-    const createdEntity2 = createComposerEntity({
-      serviceModel: containerModel.embedded_entities[0],
-      isCore: false,
-      isInEditMode: false,
-      attributes,
-      isEmbedded: true,
-      holderName: "child_container",
-    });
-
-    addDefaultEntities(graph, {
-      ...containerModel,
-      embedded_entities: [
-        {
-          ...containerModel.embedded_entities[0],
-          embedded_entities: containerModel.embedded_entities,
-        },
-      ],
-    });
-
-    expect(dispatchEventSpy).toHaveBeenCalledTimes(6);
-
-    const addedCells = graph
-      .getCells()
-      .filter((cell) => cell.get("type") !== "Link") as ServiceEntityBlock[];
-
-    //we return only top level embedded entities from addDefaultEntities so to get all we need to check graph directly
-    expect(addedCells).toHaveLength(2);
-
-    expect(addedCells[0].get("embeddedTo")).toStrictEqual(
-      createdEntity1.get("embeddedTo"),
-    );
-
-    expect(addedCells[0].get("embeddedTo")).toStrictEqual(
-      createdEntity2.get("embeddedTo"),
-    );
-  });
-});
-
-describe("populateGraphWithDefault", () => {
-  it("adds default entity for instance without embedded entities to the graph", () => {
-    const graph = new dia.Graph({});
-
-    populateGraphWithDefault(graph, parentModel);
-    const addedCells = graph.getCells() as ServiceEntityBlock[];
-
-    expect(addedCells).toHaveLength(1);
-
-    expect(addedCells[0].getName()).toStrictEqual(parentModel.name);
-  });
-
-  it("adds all required default entities for instance with embedded entities to the graph", () => {
-    const graph = new dia.Graph({});
-
-    populateGraphWithDefault(graph, containerModel);
-    const addedCells = graph
-      .getCells()
-      .filter((cell) => cell.get("type") !== "Link") as ServiceEntityBlock[];
-    const AddedLinks = graph
-      .getCells()
-      .filter((cell) => cell.get("type") === "Link") as ServiceEntityBlock[];
-
-    expect(addedCells).toHaveLength(2);
-    expect(AddedLinks).toHaveLength(1);
-
-    expect(addedCells[0].getName()).toBe(containerModel.name);
-    expect(addedCells[1].getName()).toBe(
-      containerModel.embedded_entities[0].name,
-    );
-    expect(addedCells[1].get("isEmbedded")).toBeTruthy();
-    expect(addedCells[1].get("holderName")).toBe(containerModel.name);
-    expect(addedCells[1].get("embeddedTo")).toBe(addedCells[0].id);
-  });
 });
 
 describe("addInfoIcon", () => {
