+++ src/UI/Components/Diagram/init.ts
@@ -2,11 +2,9 @@ import { RefObject } from "react";
 import { dia, shapes, ui } from "@inmanta/rappid";
 import { EmbeddedEntity, InstanceAttributeModel, ServiceModel } from "@/Core";
 import { InstanceWithRelations } from "@/Data/Managers/V2/GETTERS/GetInstanceWithRelations";
-import {
-  updateAttributes,
-  appendInstance,
-  populateGraphWithDefault,
-} from "./actions";
+import { populateGraphWithDefault } from "./actions/createMode";
+import { appendInstance } from "./actions/editMode";
+import { updateAttributes } from "./actions/general";
 import {
   applyCoordinatesToCells,
   getCellsCoordinates,
